+++ RELEASE-NOTES.md
@@ -5,6 +5,7 @@
 * Allows Arktika to read a single blob snowfield just like PoolMiner and SurfMiner
 * Change rocksdb log retention to 5 files
 * In client jsonrpc get_unspent add optional parameter to specify address to query
+* In client jsonrpc add get_block by height or hash
 
 ## 1.5.0